页面添加

+++ src/main/java/com/coosv/common/service/BaseService.java
@@ -8,7 +8,9 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.coosv.common.dao.BaseDao;
+import com.coosv.common.exception.DefaultException;
 import com.coosv.common.pojo.base.entity.BaseEntity;
+import com.coosv.common.utils.string.StringUtils;
 import com.coosv.common.web.Page;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
@@ -18,23 +20,31 @@ import com.github.pagehelper.PageInfo;
  * @author fanglei.lu
  * @date 2018年7月26日
  */
-public abstract class BaseService<D extends BaseDao<T>,T> {
+public abstract class BaseService<D extends BaseDao<T>,T extends BaseEntity> {
 	@Autowired
 	protected D dao;
 	
-	public int save(T entity) {
+	public int save(T entity)throws Exception {
+		entity.preSave();
 		return dao.save(entity);
 	}
 	
-	public int delete(String id) {
+	public int delete(String id)throws Exception {
+		if(StringUtils.isBlank(id)) {
+			throw new DefaultException("根据id删除数据请指定id");
+		}
 		return dao.delete(id);
 	}
 	
-	public int update(T entity) {
+	public int update(T entity)throws Exception {
+		entity.preUpdate();
 		return dao.update(entity);
 	}
 	
-	public T get(String id) {
+	public T get(String id)throws Exception {
+		if(StringUtils.isBlank(id)) {
+			throw new DefaultException("根据id获取数据请指定id");
+		}
 		return dao.get(id);
 	}
